Fix set-hostname under Debian Jessie.
This script is intended to be used as a dhclient exit hook, sourced (not
executed) by /sbin/dhclient-script.

The version of dhclient-script included in Debian Wheezy uses /bin/bash
as the shell to run these rather than, as documented, /bin/sh (that is,
dash).  This is fixed in Jessie, but that fix breaks this script, which
depends upon bash.

This change fixes the script to work under dash and bash equally, and
also removes the executable bit (as the script is intended to be
sourced, not executed directly, and having an executable script with a
shebang line just confuses the issue).