Fix tests when running in environment without ip in $PATH
In minimal build environments (such as those used to build Ubuntu
packages), ip will not necessarily be installed.  Its absence leads to a
warning, which causes all the self.mock_logger.warning call-count checks
to be off-by-one.

Given that the presence of ip is not important to the tests themselves,
we can safely mock it out.