Fix ip_forwarding_utils_test failures on Python 3
In Python 3, subprocess.call returns bytes (rather than str).  The mocks
returning str was causing tracebacks in a couple of places (as well as
being an incorrect test).

This updates the return values of the subprocess.Popen().communicate
mocks to use bytes, and updates the comparisons that failed as a result.