Provide an error when gsutil is not installed (#353)
Check gsutil is available when retrieving metadata scripts.

With the check we can exit early and provide a message to the user
that the executable is missing, rather than having the information about
the missing executable intermingled with a failed call thereof. Last
but not least it avoids the message that indicates that a download is
being attempted, which could lead to confusion.