Add local files to track os-login users in the event of a transient failure. (#454)
* Add a file for oslogin users to check during transient failures.

Adds a file on successful login for an oslogin user. On a backend
failure, if this file exists, we can permit that user, and deny others.

* Modify HttpGet to also set an HTTP response code.

It is not sufficient to check if CURL succeeds or not. For example, CURL
may return OK even if there is a server backend failure. Check the http
code to better check differentiate server responses.