Implements nss_cache binary for OSLogin
nss_cache_oslogin is intended to be used with the libnss-cache nss module.
This simply queries OSLogin and stores a local cache of the results,
which are then looked up by the libnss-cache module.

The cache binary is intented to be run periodically to sync with the
database. NSS should be configured to first check the cache, then fall
back on the OSLogin module if the entry is missing. This also means
account data can potentially be stale.

Note that getpwent_r() is no longer implemented for nss_oslogin. We rely
solely on the cache for this call. This is because getpwent will iterate
through all databases, and having both oslogin and cache databases
configured will cause user entries to be listed twice. The cache with a
fallback behavior only works for getpwnam/getpwuid.